more init cleanup; moved out registerRenderViewController
